Split theme.js variables into separate files to maintain code readability.
