+++ app/components/themes/theme.js
@@ -1,0 +1,14 @@
+import color from "./variables/colors";
+import components from "./variables/components";
+import font from "./variables/font";
+import layout from "./variables/layout";
+
+// A file containing all the styling variables for the styled-components theme
+const theme = {
+  color,
+  components,
+  font,
+  layout
+};
+
+export default theme;
