+++ app/components/themes/variables/colors.js
@@ -1,0 +1,31 @@
+const color = {
+  brand: {
+    primary: "#36a9e1",
+    secondary: "#2189b5",
+    tertiary: "#116892",
+    accent: ""
+  },
+  foreground: {
+    primary: "#4d4d4d",
+    secondary: "#777777"
+  },
+  background: {
+    primary: "#ffffff",
+    secondary: "#f3f3f3"
+  },
+  text: {
+    light: "#ffffff",
+    one: "#555555",
+    two: "#adadad",
+    three: "#777777"
+  },
+  system: {
+    failure: {
+      border: "#ebccd1",
+      background: "#f2dede",
+      text: "#a94442"
+    }
+  }
+};
+
+export default color;
