+++ app/components/themes/variables/components.js
@@ -1,0 +1,16 @@
+const components = {
+  paddingBaseVertical: "6px",
+  paddingBaseHorizontal: "12px",
+  paddingLargeVertical: "10px",
+  paddingLargeHorizontal: "16px",
+  paddingSmallVertical: "5px",
+  paddingSmallHorizontal: "10px",
+  paddingXsVertical: "1px",
+  paddingXsHorizontal: "5px",
+  caretWidthBase: "4px", // Width of the `border` for generating carets that indicate dropdowns.
+  caretWidthLarge: "5px", // Carets increase slightly in size for larger components
+  borderWidth: "1px",
+  spacer: "1rem"
+};
+
+export default components;
