+++ app/components/themes/variables/font.js
@@ -1,0 +1,24 @@
+const font = {
+  family: "'Montserrat', sans-serif",
+  weight: {
+    light: "300",
+    regular: "400",
+    medium: "500",
+    default: "600",
+    bold: "700"
+  },
+  size: {
+    verySmall: "0.6875rem",
+    small: "0.75rem",
+    medium: "0.875rem",
+    default: "1rem",
+    large: "1.25rem",
+    veryLarge: "3rem"
+  },
+  lineHeight: {
+    small: "1.1",
+    default: "1.5"
+  }
+};
+
+export default font;
