+++ app/components/themes/variables/layout.js
@@ -1,0 +1,10 @@
+const layout = {
+  borderRadius: {
+    small: "3px",
+    default: "5px"
+  },
+  maxWidth: "75rem",
+  padding: "0 1.25rem"
+};
+
+export default layout;
